+++ yash-semantics/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - In the `expansion::initial` module:
     - `ValueState` was renamed to `Vacancy`.
+    - `EmptyError::state` was renamed to `EmptyError::vacancy`.
 - `<expansion::Error as handle::Handle>::handle` now returns `Divert::Exit`
   instead of `Divert::Interrupt` when the `ErrExit` shell option is applicable.